requirements.txt
# requirements.txt records the full set of dependencies for development
accelerate
codetiming
datasets
dill
flash-attn
hydra-core
liger-kernel
numpy
pandas
peft
pyarrow>=19.0.0
pybind11
pylatexenc
pre-commit
ray[default]
tensordict==0.8.3
torchdata
transformers
wandb
packaging>=20.0
uvicorn
fastapi
latex2sympy2_extended
math_verify[antlr4_9_3]
vllm==0.10.0